django_api/requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.3
django_debug_toolbar==3.8.1
djangorestframework==3.15.2
drf-yasg==1.21.7
flake8==6.0.0
flake8-isort==6.0.0
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==24.1
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
